Extract sync push delta helpers
